+++ src/fhirpath.js
@@ -96,6 +96,7 @@ engine.invocationTable = {
   combine:      {fn: combining.combineFn, arity: {1: ["AnyAtRoot"]}},
   union:        {fn: combining.union,   arity: {1: ["AnyAtRoot"]}},
   intersect:    {fn: combining.intersect,   arity: {1: ["AnyAtRoot"]}},
+  exclude:      {fn: combining.exclude,   arity: {1: ["AnyAtRoot"]}},
   iif:          {fn: misc.iifMacro,    arity: {2: ["Expr", "Expr"], 3: ["Expr", "Expr", "Expr"]}},
   trace:        {fn: misc.traceFn,     arity: {1: ["String"], 2: ["String", "Expr"]}},
   toInteger:    {fn: misc.toInteger},
@@ -106,6 +107,7 @@ engine.invocationTable = {
   toTime:       {fn: misc.toTime},
   toBoolean:    {fn: misc.toBoolean},
   toQuantity:   {fn: misc.toQuantity, arity: {0: [], 1: ["String"]}},
+  hasValue:     {fn: misc.hasValueFn},
   convertsToBoolean:    {fn: misc.createConvertsToFn(misc.toBoolean, 'boolean')},
   convertsToInteger:    {fn: misc.createConvertsToFn(misc.toInteger, 'number')},
   convertsToDecimal:    {fn: misc.createConvertsToFn(misc.toDecimal, 'number')},
@@ -433,13 +435,13 @@ engine.realizeParams = function(ctx, parentData, args) {
 function makeParam(ctx, parentData, type, param) {
   if(type === "Expr"){
     return function(data) {
-      ctx.$this = util.arraify(data);
-      return engine.doEval(ctx, ctx.$this, param);
+      const $this = util.arraify(data);
+      return engine.doEval({ ...ctx, $this }, $this, param);
     };
   }
   if(type === "AnyAtRoot"){
-    ctx.$this = ctx.$this || ctx.dataRoot;
-    return engine.doEval(ctx, ctx.$this, param);
+    const $this = ctx.$this || ctx.dataRoot;
+    return engine.doEval({ ...ctx, $this}, $this, param);
   }
   if(type === "Identifier"){
     if(param.type === "TermExpression") {
@@ -453,7 +455,6 @@ function makeParam(ctx, parentData, type, param) {
     return engine.TypeSpecifier(ctx, parentData, param);
   }
 
-  ctx.$this = parentData;
   const res = engine.doEval(ctx, parentData, param);
   if(type === "Any") {
     return res;
@@ -469,7 +470,7 @@ function makeParam(ctx, parentData, type, param) {
 }
 
 function doInvoke(ctx, fnName, data, rawParams){
-  var invoc = engine.invocationTable[fnName];
+  var invoc = ctx.userInvocationTable?.[fnName] ?? engine.invocationTable[fnName];
   var res;
   if(invoc) {
     if(!invoc.arity){
@@ -647,13 +648,15 @@ function parse(path) {
  * @param {(object|object[])} resource -  FHIR resource, bundle as js object or array of resources
  *  This resource will be modified by this function to add type information.
  * @param {object} parsedPath - a special object created by the parser that describes the structure of a fhirpath expression.
- * @param {object} [context] - a hash of variable name/value pairs.
- * @param {object} [model] - The "model" data object specific to a domain, e.g. R4.
+ * @param {object} context - a hash of variable name/value pairs.
+ * @param {object} model - The "model" data object specific to a domain, e.g. R4.
  *  For example, you could pass in the result of require("fhirpath/fhir-context/r4");
- * @param {object} [options] - additional options:
+ * @param {object} options - additional options:
  * @param {boolean} [options.resolveInternalTypes] - whether values of internal
- * @param {function} [options.traceFn] - An optional trace function to call when tracing
  *  types should be converted to strings, true by default.
+ * @param {function} [options.traceFn] - An optional trace function to call when tracing.
+ * @param {object} [options.userInvocationTable] - a user invocation table used
+ *  to replace any existing or define new functions.
  */
 function applyParsedPath(resource, parsedPath, context, model, options) {
   constants.reset();
@@ -677,8 +680,11 @@ function applyParsedPath(resource, parsedPath, context, model, options) {
     }, {});
   }
   let ctx = {dataRoot, vars: Object.assign(vars, context), model};
-  if (options && options.traceFn) {
+  if (options.traceFn) {
     ctx.customTraceFn = options.traceFn;
+  }
+  if (options.userInvocationTable) {
+    ctx.userInvocationTable = options.userInvocationTable;
   }
   return  engine.doEval(ctx, dataRoot, parsedPath.children[0])
     // engine.doEval returns array of "ResourceNode" and/or "FP_Type" instances.
@@ -744,10 +750,12 @@ function resolveInternalTypes(val) {
  *  For example, you could pass in the result of require("fhirpath/fhir-context/r4");
  * @param {object} [options] - additional options:
  * @param {boolean} [options.resolveInternalTypes] - whether values of internal
- * @param {function} [options.traceFn] - An optional trace function to call when tracing
  *  types should be converted to standard JavaScript types (true by default).
  *  If false is passed, this conversion can be done later by calling
  *  resolveInternalTypes().
+ * @param {function} [options.traceFn] - An optional trace function to call when tracing.
+ * @param {object} [options.userInvocationTable] - a user invocation table used
+ *  to replace any existing or define new functions.
  */
 function evaluate(fhirData, path, context, model, options) {
   return compile(path, model, options)(fhirData, context);
@@ -767,14 +775,41 @@ function evaluate(fhirData, path, context, model, options) {
  *  For example, you could pass in the result of require("fhirpath/fhir-context/r4");
  * @param {object} [options] - additional options:
  * @param {boolean} [options.resolveInternalTypes] - whether values of internal
- * @param {function} [options.traceFn] - An optional trace function to call when tracing
  *  types should be converted to strings, true by default.
+ * @param {function} [options.traceFn] - An optional trace function to call when tracing.
+ * @param {object} [options.userInvocationTable] - a user invocation table used
+ *  to replace any existing or define new functions.
  */
 function compile(path, model, options) {
   options = {
     resolveInternalTypes: true,
     ... options
   };
+
+  const userInvocationTable = options.userInvocationTable;
+  if (userInvocationTable) {
+    options.userInvocationTable = Object.keys(userInvocationTable).reduce(
+      (invocationTable, fnName) => {
+        if (userInvocationTable[fnName].internalStructures) {
+          invocationTable[fnName] = userInvocationTable[fnName];
+        } else {
+          invocationTable[fnName] = {
+            ...userInvocationTable[fnName],
+            fn: (...args) => {
+              return userInvocationTable[fnName].fn.apply(
+                // When we check Array.isArray(arg), we are checking if the
+                // singleton function has been called. An alternative to this is
+                // to check that the type of the argument is Integer, Boolean,
+                // Number, or String.
+                this, args.map(arg => Array.isArray(arg) ? arg.map(item => util.valData(item)) : arg)
+              );
+            }
+          };
+        }
+        return invocationTable;
+      }, {});
+  }
+
   if (typeof path === 'object') {
     const node = parse(path.expression);
     return function (fhirData, context) {
@@ -814,5 +849,7 @@ module.exports = {
   resolveInternalTypes,
   types: typesFn,
   // Might as well export the UCUM library, since we are using it.
-  ucumUtils: require('@lhncbc/ucum-lhc').UcumLhcUtils.getInstance()
+  ucumUtils: require('@lhncbc/ucum-lhc').UcumLhcUtils.getInstance(),
+  // Utility functions that can be used to implement custom functions
+  util
 };